+++ src/models/components/llmconfig.ts
@@ -4,15 +4,26 @@
 
 import * as z from "zod";
 import { remap as remap$ } from "../../lib/primitives.js";
-import {
-  Provider,
-  Provider$inboundSchema,
-  Provider$outboundSchema,
-} from "./provider.js";
+import { safeParse } from "../../lib/schemas.js";
+import { ClosedEnum } from "../../types/enums.js";
+import { Result as SafeParseResult } from "../../types/fp.js";
+import { SDKValidationError } from "../errors/sdkvalidationerror.js";
+
+/**
+ * Provider of the LLM model.
+ */
+export const Provider = {
+  AzureOpenai: "azure_openai",
+  Openai: "openai",
+} as const;
+/**
+ * Provider of the LLM model.
+ */
+export type Provider = ClosedEnum<typeof Provider>;
 
 export type LlmConfig = {
   /**
-   * LLM API provider.
+   * Provider of the LLM model.
    */
   provider?: Provider | undefined;
   /**
@@ -30,12 +41,31 @@ export type LlmConfig = {
 };
 
 /** @internal */
+export const Provider$inboundSchema: z.ZodNativeEnum<typeof Provider> = z
+  .nativeEnum(Provider);
+
+/** @internal */
+export const Provider$outboundSchema: z.ZodNativeEnum<typeof Provider> =
+  Provider$inboundSchema;
+
+/**
+ * @internal
+ * @deprecated This namespace will be removed in future versions. Use schemas and types that are exported directly from this module.
+ */
+export namespace Provider$ {
+  /** @deprecated use `Provider$inboundSchema` instead. */
+  export const inboundSchema = Provider$inboundSchema;
+  /** @deprecated use `Provider$outboundSchema` instead. */
+  export const outboundSchema = Provider$outboundSchema;
+}
+
+/** @internal */
 export const LlmConfig$inboundSchema: z.ZodType<
   LlmConfig,
   z.ZodTypeDef,
   unknown
 > = z.object({
-  provider: Provider$inboundSchema.optional(),
+  provider: Provider$inboundSchema.default("azure_openai"),
   model: z.string().default("gpt-4o"),
   version: z.nullable(z.string()).optional(),
   api_version: z.nullable(z.string()).optional(),
@@ -47,7 +77,7 @@ export const LlmConfig$inboundSchema: z.ZodType<
 
 /** @internal */
 export type LlmConfig$Outbound = {
-  provider?: string | undefined;
+  provider: string;
   model: string;
   version?: string | null | undefined;
   api_version?: string | null | undefined;
@@ -59,7 +89,7 @@ export const LlmConfig$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   LlmConfig
 > = z.object({
-  provider: Provider$outboundSchema.optional(),
+  provider: Provider$outboundSchema.default("azure_openai"),
   model: z.string().default("gpt-4o"),
   version: z.nullable(z.string()).optional(),
   apiVersion: z.nullable(z.string()).optional(),
@@ -80,4 +110,18 @@ export namespace LlmConfig$ {
   export const outboundSchema = LlmConfig$outboundSchema;
   /** @deprecated use `LlmConfig$Outbound` instead. */
   export type Outbound = LlmConfig$Outbound;
+}
+
+export function llmConfigToJSON(llmConfig: LlmConfig): string {
+  return JSON.stringify(LlmConfig$outboundSchema.parse(llmConfig));
+}
+
+export function llmConfigFromJSON(
+  jsonString: string,
+): SafeParseResult<LlmConfig, SDKValidationError> {
+  return safeParse(
+    jsonString,
+    (x) => LlmConfig$inboundSchema.parse(JSON.parse(x)),
+    `Failed to parse 'LlmConfig' from JSON`,
+  );
 }