+++ RELEASES.md
@@ -71,3 +71,13 @@ Based on:
 - [typescript v0.1.0-alpha.3] .
 ### Releases
 - [NPM v0.1.0-alpha.3] https://www.npmjs.com/package/syllable-sdk/v/0.1.0-alpha.3 - .
+
+## 2024-11-18 15:59:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.440.1 (2.460.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.1.0-alpha.4] .
+### Releases
+- [NPM v0.1.0-alpha.4] https://www.npmjs.com/package/syllable-sdk/v/0.1.0-alpha.4 - .